Struct google_appstate1::StateMethods

pub struct StateMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on state resources. It is not used directly, but through the AppState hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_appstate1 as appstate1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use appstate1::AppState;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AppState::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `clear(...)`, `delete(...)`, `get(...)`, `list(...)` and `update(...)`
// to build up your call.
let rb = hub.states();

Methods

impl<'a, C, A> StateMethods<'a, C, A>

fn delete(&self, state_key: i32) -> StateDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a key and the data associated with it. The key is removed and no longer counts against the key quota. Note that since this method is not safe in the face of concurrent modifications, it should only be used for development and testing purposes. Invoking this method in shipping code can result in data loss and data corruption.

Arguments

stateKey - The key for the data to be retrieved.

fn get(&self, state_key: i32) -> StateGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the data corresponding to the passed key. If the key does not exist on the server, an HTTP 404 will be returned.

Arguments

stateKey - The key for the data to be retrieved.

fn clear(&self, state_key: i32) -> StateClearCall<'a, C, A>

Create a builder to help you perform the following task:

Clears (sets to empty) the data for the passed key if and only if the passed version matches the currently stored version. This method results in a conflict error on version mismatch.

Arguments

stateKey - The key for the data to be retrieved.

fn list(&self) -> StateListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all the states keys, and optionally the state data.

fn update(&self, request: UpdateRequest, state_key: i32) -> StateUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Update the data associated with the input key if and only if the passed version matches the currently stored version. This method is safe in the face of concurrent writes. Maximum per-key size is 128KB.

Arguments

request - No description provided.

stateKey - The key for the data to be retrieved.